Add GNUPG test setup
For a basic setup for the test environment a gpg1 keyring and
keypair is added and the keys are migrated to gpg 2.1 - so the setup
will work both for gpg1 and gpg2.1.  If any of the gpg features is
needed in the specs, ENV["GNUPGHOME"] has to be set to 'spec/gnupg' using the
`set_test_gnupg_home` method that is added to the spec helper.

The setup is used in the spec to test the key method of the list
model.