Use a temp gpghome with a short path in test env
With the previous setup the tests would fail in a LXC environment
because of the deep folder hierarchy. The gpg agent does not like long
path names and will fail stating that the socket name 'very/long/path'
is too long.

We now use a tmpdir as gpghome with a shorter path.

Additionally the setup of the gpghome was moved to the before suite hook
in the spec helper. Clean up is done in the after suite hook. It deletes
all tmpfiles, sets ENV["GNUPGHOME"] to nil and kills the gpg-agent.