Don't enable delivery for admins that weren't a member in v2
In v2 admins & members were 2 seperate arrays and hence it could be
totally fine to be an admin without receiving mails on the list.

In v3 we do not make that disctinction anymore, rather admins are
normal subscriptions with an admin flag indicating their status.
Also there is now a delivery_enabled flag that can be set to false
on a subscription.

When migrating a v2 list to v3 we didn't respect that case, when
admins were actually not members, but just subscribed them.

Closes #213