refactor Schleuder::Filters::Runner
This is the base work for future improvements of the filters runner,
but is not changing any functionality.

We make the runner stateful, as it runs for one list, but might run
multiple times. However, the mail object that the filters operate on
are not part of the state, as the mail might be (ex-)changed by
filters or other codeflow in between the multiple executions of
the filter.

Also we add tests for the runner, to cover its basic functionality.
This also discovered a bug, where bounces_drop_on_headers
would not work as intended, as the headers of the mail were not
properly checked, which is also fixed as part of that rework.