migrate message fixing to filters
based on the work we did in the previous commits to make the filters
also run before the setup method is being called, we are now able
to extract methods that fix broken/invalid messages into the filters
engine. This reduces the amount of code in `Mail::Message`, as well
as isolates the message fixing code.

To prove that we do not break the end-to-end experience, we also add
integration tests for these use-cases, as well as migrate the previous
message related tests to their own dedicated filter tests.

This is based on the discussion that followed the proposal in #234