fix smtp daemon when previous test got aborted
If the previous test was aborted, the daemon might still run or
at least the directory still exists and we will try to stop it.

Stopping it however, deletes the directory to which we want to
write the pid file once we started it. So we must inverse the
order on how to deal with it.

Part of #235