Fix exceptions from migrations for some setups.
Some commits contain changes to db/schema.rb but don't include the
updated migration-timestamp. This led to the situation in which setups
that started in a certain time-frame want to run some migrations,
which caused exceptions. This patch fixes that situation so the
problematic setups can run all migrations successfully and "catch up".