Fix handling protected subject.
Enigmail sends a "protected subject" since v2.0, which leaked due to the
way mail-gpg copies headers from and to mime-parts.
Also since the first mime-part in those messages is always a "protected
headers" part, request-messages were bounced as empty because we only
look into the first mime part.

This fix still has one shortcoming: the real, "protected" subject is not
included into the primary mime-headers of the encrypted content. It is
only contained in the "protected headers" mime-part which people get to
read as part of the body. We need a change in mail-gpg to fix this.