CONTRIBUTING: Add note how to enable shared runners to run CI tests
The default, shared 0xacab runner is quite slow, therefore, it's
disabled in our repo. If people forking our repo, the CI / CD settings
are copied as well. This makes the CI jobs for them to be stuck up until
GitLab kills them because of a timeout. Improve this situation at least
via extending the contributing guide with a note regarding this.

Relates #308