fix #309 - be more robust when dirmngr fails while refreshing keys
It can happen that a single failure in dirmngr makes all
subsequent key refreshes of that run also failing. E.g. as it
marked the keyserver internally as dead, although it might not be.
See #309 for background.

By killing dirmngr on any error that we don't filter out anyways,
we make sure that any subsequent run won't inherit any problem
that might persist over a dirmngr session. E.g. this helps with
the most prominent case, when dirmngr marks a onionservice keyserver
as dead, although it was just the tor connection that had once a
problem and subsequent runs will succeed.