CI: Fix APT caching, once again
It seems the code introduced in 18d2b528 still didn't work all the time.
By now, I believe the GitLab CI caching to be buggy.

For now, just rely on the default cache key, and run 'apt-get update'
everytime. Especially the later part is not optimal, but for now it
seems it's the way to go, unfortunately.

Relates #316