debian/schleuder.postinst: Fix permissions of files and directories
A user reported upstream, that the files inside '/etc/schleuder' had the
executable bit set.

At the same time it was reported in bugs #889060 and #889488, that doing
recursive chmod or chmod might lead to root escalation on systems with
fs.protected_hardlinks=0.

This commit fixes both problems via using 'find'. Additionally, the
permissions are made more strict, to ensure that the direcotories and
files are only accessible by the schleuder user. Before, the schleuder
group was allowed access, too.