debian/{rules,s-a-d.service}: Use new dh_installsystemd helper
In debhelper compat level 11, the systemd sequence for dh was removed,
and the dh_systemd_{enable,start,stop} helpers as well. Accordingly,
drop the obsolete helpers and make use of the new dh_installsystemd
helper.

Moreover, if specifying --name for dh_installsystemd, debhelper now
expects the systemd service file to be installed to be named as
$package-name.$service-name.service. Accordingly, rename the service
file.