Improve parsing of keywords - first part of #249
This changes the way how we parse for keywords to the following
approach:

  1. Ignore any empty/pure-whitespace lines at the beginning of
     the first plain_text_part message.
  2. If you match something /^X\-.*/ start matching for plugins,
     else abort.
  3. Once you matched a plugin, but you have a new
     non-empty/pure-whitespace line starting with something
     different than /^X\-.*/ abort looking for plugin names.

This fixes the most common problems with lines starting with `X-`
further down in the body of an email, while keeping all of the
existing behavior and functionality.

Any further improvements to the code base and the parsing of
mails as discussed in #249 is postponed to future work. Tis work
should go together with work in #359 which will allow to change
the structure of plugins.