CI: Introduce job to build Debian package
This enables us to get fast(er) feedback in merge requests regarding the
Debian-side of things. It allows to spot errors and to find breaking
changes earlier.

Besides, this will reduce the time between the upstream release and the
upload to Debian. It enables to do necessary Debian packaging changes
while the new upstream version is still in development, in contrast to
postpone if after the actual release, which was the situation before
this commit.

Less pain, more joy to come. \m/

Relates #324