Fix #360 - do not allow setting an empty fingerprint through `SET-FINGERPRINT`
Check for a valid fingerprint as argument and do not accept an empty
fingerprint. This will not anymore allow to unset a fingerprint
through the `SET-FINGERPRINT`. This functionality will be superseeded
by another keyword.

As part of that fix, wie centralize checking for a valid fingerpint
and constrain the check to be either 32 (v3) or 40 (v4) characters
long.