Fix spec for factory-bot 5.0.
They changed a default setting so now associated resources are not
created but also only built when using build(), which means the
associated resources don't have an ID and thus are not valid.

Rubies 2.1 and 2.2 use v4.1 because they are incompatible with v5.0, but
all later versions of ruby use v5.0.