CI: Enforce rubocop, disallow errors
We've introduced rubocop some time ago, to help with code linting. We
still don't like the name, but it might improve code quality in the
future. Due to some errors in the beginning, the results weren't
reliable. This seems fixed since a while, therefore, disallow errors and
close the corresponding issue.

Closes #382