CI: Add job to check spec filenames for trailing '_spec' part
Some time ago we added some spec files, however, the filenames were
missing the trailing '_spec' part, so these were never executed.

This commit adds a new CI job to check for this problem and to catch it
earlier next time. For now, this job allows errors. Lets revisit this in
the future to probably make the result fatal, that is, block a merge
request based on the result.

This job is limited to changes made in the relevant spec directories,
however, due to the nature GitLab works currently, if pushing new
branches, this job will be executed regardless of which files were
changed. However, as this job is quite "light" regarding resources, this
shouldn't be of any concern.

Some tests were done via intermediate commits to ensure this works as
expected.

Relates !284