CI: Deploy code coverage overview as HTML to make it actually usable
Code coverage was added to the CI some time ago. Up until now, a part of
the result was printed at the end of the rspec test jobs, accessible via
the individual job logs, barely usable.

This commit introduces deploying the HTML result and making it
accessible via a dedicated page, allowing to search within and filter
the data, in a much more convenient and appealing way.

While I'm not hunting for 100 percent code coverage, I think having at
least a better insight into this topic is quite valuable. I hope this
leads to a better understanding of our test suite, possibly identifying
weak points (which then can be improved) and in the end to a better code
base.