CI: codespell: ignore code coverage dir leading to false positives
Due to the introduced artifacts to store the code coverage report, and
to "hand it over" from one stage to the next, these artifacts are now
downloaded as well before the codespell job. This lead to false
positives, spelling errors within this directory, we're not really
interested in. Therefore, ignore this directory if running codespell.