fix #457 - ensure proper encoding if no charset is set on a part
If a part did not have a charset set, we falled back to US-ASCII
and then tried to encode the string into that. However, this only
worked if the part had actually only ascii charcters, but as soon
it contains non-ascii characters we failed. With this change we
are encoding such parts as utf-8, thus adopting the future.