specs: fix errors on IPv6-only machines
Do not rely on '127.0.0.1', but on 'localhost'.

dirmngr receives some extra care, so it's able to cope with it.

gpgconf --kill might hang indefinitely, therefore, rely on pkill.

Relax the expected output of gpg if refreshing keys without a keyserver
being available, as gpg might report various errors in such a situation.

The dependency on dirmngr will be dropped soon, anyway, so I deem this
acceptable as a workaround, for now.

Closes #472