Use class variable instead of constant.
Feels more logical to specify the keyword and the arguments together.

Also constants are specific to the class their are defined in, so to
access the keyword-handler's constant from the base class one has to
call `self.class::WANTED_ARGUMENTS`, which I dislike.
And with the new way there's no need to check if the arguments have been
specified, becaus they are a mandatory method-argument.