Add option to use the sender's email in headers
This configurable option is introduced to tackle the following usability problem:

Previously:
When a user receives an email from the list, the FROM-header and
REPLY-TO-header were set to the list address. With that the mail client
will automatically set the REPLY-TO address to the list address.
Since this might confuse some users, we wanted to make it configurable
to send the original's senders address in the REPLY-TO/FROM-header.

This commit:
An option "use_unencrypted_sender_addresses_in_header" is introduced.
When enabled, the FROM-header is set to a munched version of the
original sender's address and the REPLY-TO header is set to the original sender's address.
This option is disabled by default, since enabling it means the sender's
unencrypted address will be sent to everyone on the list.
In some use cases where usability is more important than the impact
of such "leakage", enabling can be considered by using:

schleuder-cli lists set <listname> use_unencrypted_sender_addresses_in_header true