CI: changelog: fetch our repository
The changelog job requires to fetch the latest state of the repository,
otherwise the result of this check might not be correct.

Up until now, 'origin' was fetched, which points, by default, to the
remote repository a local checkout was cloned from, initially.

This works for our repository, but might fail for forks, depending on
when they were forked and if they were updated in the meantime.

This commit ensures the correct repository is fetched. Besides this, the
amount of data fetched is limited.

Ref https://0xacab.org/schleuder/schleuder/-/merge_requests/374