Incorporate stripped down and lightly integrated code of Mail::Gpg
We experienced unpleasent surprises and patched the code anyway, so we
now chose to incorporate it (it's licensed under "MIT", which allows
that without any obligations).
The code is stripped from the parts we don't use, changed in those parts
we previously patched (Mail::Gpg#encrypt(),
Mail::Gpg::SignPart#verify_signature()), and a little integrated
(code of MessagePatch pulled into our own Mail::Message).

Beyond that, the code is verbatim from `mail-gpg` version 0.4.2.

I don't want to integrate it further because we might want to get rid of
gpg in a future moment.