specs: make it possible to conditionally require 'byebug'
This helps in downstream environments, like Debian, via relying on
environment variables, instead of patches, to not require 'byebug'.

Similar code is already in place in regards to the usage of 'bundler'
and 'simplecov'.

This commit doesn't include a CHANGELOG note, as it's not really
relevant to most users, I believe.