Use constant to specify which routes should be public
I tried for hours to find a solution using before-filters and
conditions, but all options either didn't work for all routes (e.g. a
before-filter in a namespace failed for `get ".json"`), or were based on
hacks or internal APIs of sinatra.
So I took a step back and used the simple and booring solution.